Renamed cache.rs to redis.rs
