+++ src/utils/context.rs
@@ -3,7 +3,7 @@ use rocket::request::{self, FromRequest, Request};
 use rocket::Outcome;
 use rocket::State;
 
-use crate::cache::cache::ServiceCache;
+use crate::cache::redis::ServiceCache;
 use crate::cache::Cache;
 use crate::config::scheme;
